+++ Source/Utils/niLicense/niLicense.cpp
@@ -119,14 +119,14 @@ int main(int argc, char* argv[])
 		nRetVal = xnOSStrNCopy(license.strVendor, strVendor, strlen(strVendor) + 1, sizeof(license.strVendor));
 		if (nRetVal != XN_STATUS_OK)
 		{
-			printf("Error: Vendor name is too long (should be up to %u characters).", sizeof(license.strVendor) - 1);
+			printf("Error: Vendor name is too long (should be up to %lu characters).", (unsigned long) (sizeof(license.strVendor) - 1));
 			return -1;
 		}
 
 		nRetVal = xnOSStrNCopy(license.strKey, strKey, strlen(strKey) + 1, sizeof(license.strKey));
 		if (nRetVal != XN_STATUS_OK)
 		{
-			printf("Error: Key is too long (should be up to %u characters).", sizeof(license.strKey) - 1);
+			printf("Error: Key is too long (should be up to %lu characters).", (unsigned long) (sizeof(license.strKey) - 1));
 			return -1;
 		}
 	}
